+++ src/main/java/com/sda/service/AdService.java
@@ -7,7 +7,6 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 
 import java.util.List;
-import java.util.Optional;
 
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 public class AdService {
@@ -28,7 +27,7 @@ public class AdService {
         adRepository.save(ad);
     }
 
-    public List<Ad> getAllAdsFiltered(Double minPrice, Double maxPrice, Integer minMileage, Integer maxMileage, Integer minYear, Integer maxYear, String company, String sort){
+    public List<Ad> getAllAdsFiltered(Integer minPrice, Integer maxPrice, Integer minMileage, Integer maxMileage, Integer minYear, Integer maxYear, String company, String sort){
         return adRepository.getAdsFiltered(minPrice, maxPrice, minMileage, maxMileage, minYear, maxYear, company, sort);
     }
     public List<Ad> getAdsByUser(User user){
